# apps/api/security/openapi.json
{
  "components": {
    "schemas": {
      "ContextEnvelope": {
        "properties": {
          "context": {
            "$ref": "#/components/schemas/EffectiveContext"
          }
        },
        "required": [
          "context"
        ],
        "type": "object"
      },
...
      "EffectiveContext": {
        "description": "The resolved permissions and ceilings for a user at a point in time. This\nstruct is serialized as the `/auth/me` envelope, so its field names are\npart of the public API contract.",
        "properties": {
          "allowed_template_ids": {
            "description": "Union of every member group's whitelist (group-only authorization). No\npersonal whitelist and no creator self-whitelist (spec Decision 4/5).\nHidden templates are always excluded — they never appear in the\nwhitelist, so clients can treat this list as \"everything I may launch\".",
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "type": "array"
          },
          "can_create_template": {
            "type": "boolean"
          },
          "can_manage_docker": {
            "type": "boolean"
          },
          "can_manage_group_instances": {
            "type": "boolean"
          },
          "can_manage_registry": {
            "type": "boolean"
          },
          "can_manage_users": {
            "type": "boolean"
          },
          "direct_max_instances": {
            "description": "The personal override, if any. The effective ceiling is the max of this\nand the group maxima, so a personal ceiling can only raise, never lower,\nwhat the groups grant.",
            "format": "int32",
            "type": [
              "integer",
              "null"
            ]
          },
          "effective_max_instances": {
            "description": "`0` means \"no ceiling\" (matches the `host_instance_limit = 0`\nconvention). Non-zero values are exact per-user limits. Resolved as the\n*maximum* of the personal ceiling and every group ceiling, where 0/NULL\n= unlimited is the highest (spec Decision 4).",
            "format": "int32",
            "type": "integer"
          },
          "group_ids": {
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "type": "array"
          },
          "is_admin": {
            "description": "Root: derived from Admin-group membership, never stored (spec Decision\n4). Replaces the dropped `is_system_admin` column.",
            "type": "boolean"
          },
          "tier": {
            "description": "The maximum tier across the user's group memberships (0/1/2).",
            "format": "int32",
            "type": "integer"
          },
          "user_id": {
            "format": "uuid",
            "type": "string"
          },
          "username": {
            "type": "string"
          }
        },
        "required": [
          "user_id",
          "username",
          "is_admin",
          "tier",
          "can_create_template",
          "can_manage_users",
          "can_manage_group_instances",
          "can_manage_docker",
          "can_manage_registry",
          "effective_max_instances",
          "allowed_template_ids",
          "group_ids"
        ],
        "type": "object"
      },
      "GroupListEnvelope": {
        "properties": {
          "groups": {
            "items": {
              "$ref": "#/components/schemas/GroupSchema"
            },
            "type": "array"
          }
        },
        "required": [
          "groups"
        ],
        "type": "object"
      },
      "GroupSchema": {
        "properties": {
          "can_create_template": {
            "type": "boolean"
          },
          "can_manage_docker": {
            "type": "boolean"
          },
          "can_manage_group_instances": {
            "type": "boolean"
          },
          "can_manage_registry": {
            "type": "boolean"
          },
          "can_manage_users": {
            "type": "boolean"
          },
          "description": {
            "type": [
              "string",
              "null"
            ]
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "kind": {
            "type": [
              "string",
              "null"
            ]
          },
          "max_instances": {
            "format": "int32",
            "type": [
              "integer",
              "null"
            ]
          },
          "name": {
            "type": "string"
          },
          "template_ids": {
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "type": "array"
          }
        },
        "required": [
          "id",
          "name",
          "can_create_template",
          "can_manage_users",
          "can_manage_group_instances",
          "can_manage_docker",
          "can_manage_registry",
          "template_ids"
        ],
        "type": "object"
      },
...
      "LoginRequest": {
        "properties": {
          "password": {
            "type": "string"
          },
          "username": {
            "type": "string"
          }
        },
        "required": [
          "username",
          "password"
        ],
        "type": "object"
      },
...
      "SettingsEnvelope": {
        "properties": {
          "settings": {
            "$ref": "#/components/schemas/SystemSettings"
          }
        },
        "required": [
          "settings"
        ],
        "type": "object"
      },
      "SystemSettings": {
        "description": "The single global-policy knob exposed by the admin settings API: the host\ninstance ceiling (`0` = unlimited). The host-capacity / shared-fuse fields\nwere dropped with the old quota pipeline.",
        "properties": {
          "host_instance_limit": {
            "format": "int32",
            "type": "integer"
          }
        },
        "required": [
          "host_instance_limit"
        ],
        "type": "object"
      },
...
      "ValidateEnvelope": {
        "properties": {
          "is_admin": {
            "type": "boolean"
          },
          "tier": {
            "format": "int32",
            "type": "integer"
          },
          "user_id": {
            "format": "uuid",
            "type": "string"
          },
          "username": {
            "type": "string"
          }
        },
        "required": [
          "user_id",
          "username",
          "is_admin",
          "tier"
        ],
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "OpenAPI spec for the 17 security-fuzzable endpoints (see .scratch/security-fuzzing). The 8 admin-gated operations are tagged `admin-gated` for the low-privilege RBAC-boundary pass. The spec is export-only and never served.",
    "license": {
      "name": ""
    },
    "title": "OpenWorkspace API — security fuzz surface",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/admin/settings": {
      "get": {
        "operationId": "get_settings",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SettingsEnvelope"
                }
              }
            },
            "description": "global policy settings"
          },
          "401": {
            "description": "missing or invalid ow_token"
          },
          "403": {
            "description": "requires admin"
          },
          "500": {
            "description": "internal server error"
          }
        },
        "tags": [
          "admin-gated"
        ]
      }
    },
    "/api/auth/login": {
      "post": {
        "operationId": "login",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContextEnvelope"
                }
              }
            },
            "description": "authenticated; sets the ow_token cookie"
          },
          "400": {
            "description": "request body is not valid JSON (syntax error)"
          },
          "401": {
            "description": "invalid credentials"
          },
          "415": {
            "description": "missing Content-Type: application/json"
          },
          "422": {
            "description": "malformed JSON body"
          },
          "500": {
            "description": "internal server error"
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/me": {
      "get": {
        "operationId": "me",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContextEnvelope"
                }
              }
            },
            "description": "current effective context"
          },
          "401": {
            "description": "missing or invalid ow_token"
          },
          "500": {
            "description": "internal server error"
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/validate": {
      "get": {
        "operationId": "validate",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidateEnvelope"
                }
              }
            },
            "description": "session identity"
          },
          "401": {
            "description": "missing or invalid ow_token"
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
...
    "/api/groups": {
      "get": {
        "operationId": "list_groups",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GroupListEnvelope"
                }
              }
            },
            "description": "group catalog with template whitelists"
          },
          "401": {
            "description": "missing or invalid ow_token"
          },
          "403": {
            "description": "requires can_manage_users or admin"
          },
          "500": {
            "description": "internal server error"
          }
        },
        "tags": [
          "admin-gated"
        ]
      }
    },
...
  }
}